Convert autocomplete tests to TS.
